:recycle:refactor: 자랑하기 상세페이지 컴포넌트 분리
